Move and rename some tests
